Introduce websocket rooms
This introduces the idea of having "rooms" of clients. The idea is that
various different actions can be broadcast to all of the clients. Such
actions may include pausing, playing, skipping to the next song, seeking
within the current song.

This helps with the implementation of synchronising multiple players.
(Issue #8)